TESB-22605 Refactoring after OAS tools move from TESB to TDI
