+++ main/plugins/org.talend.designer.esb.components.rs.provider/components/tRESTRequest/tRESTRequest_java.xml
@@ -31,8 +31,13 @@
 
 	<PARAMETERS>
 	
-		<PARAMETER NAME="OAS_IMPORT" FIELD="EXTERNAL" NUM_ROW="1" GROUP="OAS_GROUP"  SHOW_IF="IS_STUDIO_EE_VERSION == 'true'" />
-		
+		<PARAMETER NAME="PROPERTY" FIELD="PROPERTY_TYPE" GROUP="OAS_GROUP" SHOW="true"
+			NUM_ROW="1" REPOSITORY_VALUE="METADATA_DATASERVICES" />
+
+		<!-- 
+		<PARAMETER NAME="OAS_IMPORT" FIELD="EXTERNAL" NUM_ROW="1" GROUP="OAS_GROUP" SHOW_IF="IS_STUDIO_EE_VERSION == 'true'" />
+		-->
+	
 		<PARAMETER NAME="SCHEMA_WRONG_CALLS" FIELD="SCHEMA_TYPE" NUM_ROW="2" CONTEXT="WRONG_CALLS" SHOW="false">
 			<TABLE READONLY="true">
 				<COLUMN NAME="uri" TYPE="id_String" LENGTH="255" READONLY="true"/>
