+++ main/plugins/org.talend.designer.esb.components.ws.tools.extensions/src/org/talend/designer/esb/components/ws/tools/extensions/external/IOASDecoder.java
@@ -26,6 +26,8 @@ public interface IOASDecoder {
 
     public String getEndpoint();
 
+    public String getName();
+
     public String getDocumentationComment();
 
     public Map<String, RestAPIMapping> getMappings();
